requirements.txt
asgiref==3.9.2
async-timeout==5.0.1
certifi==2025.10.5
charset-normalizer==3.4.3
Django==4.2.25
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
packaging==25.0
PyMySQL==1.1.2
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.16
